+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/PartyService.cs
@@ -33,6 +33,7 @@ using Stormancer.Server.Plugins.Users;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Reactive.Linq;
 using System.Reactive.Threading.Tasks;
@@ -101,7 +102,6 @@ namespace Stormancer.Server.Plugins.Party
             _stormancerPartyPlatformSupport = stormancerPartyPlatformSupport;
             this.invitationCodes = invitationCodes;
             ApplySettings(configuration.Settings);
-
         }
 
         private const string JoinDeniedError = "party.joinDenied";
@@ -166,7 +166,7 @@ namespace Stormancer.Server.Plugins.Party
 
         internal async Task OnConnecting(IScenePeerClient peer)
         {
-            var handlers = _handlers;
+            var handlers = _handlers();
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 //Todo jojo later
@@ -191,7 +191,7 @@ namespace Stormancer.Server.Plugins.Party
                 }
 
                 var ctx = new JoiningPartyContext(this, session, _partyState.PendingAcceptedPeers.Count + _partyState.PartyMembers.Count);
-                await handlers().RunEventHandler(h => h.OnJoining(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnJoining", ex));
+                await handlers.RunEventHandler(h => h.OnJoining(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnJoining", ex));
                 if (!ctx.Accept)
                 {
                     Log(LogLevel.Trace, "OnConnecting", "Join denied by event handler", peer.SessionId, session.User?.Id);
@@ -342,7 +342,6 @@ namespace Stormancer.Server.Plugins.Party
 
         public async Task UpdateSettings(PartySettingsDto partySettingsDto, CancellationToken ct)
         {
-            var handlers = _handlers;
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 if (ct.IsCancellationRequested)
@@ -354,9 +353,14 @@ namespace Stormancer.Server.Plugins.Party
                 {
                     throw new ClientException(GameFinderNameError);
                 }
+
+                await using var scope = _scene.CreateRequestScope();
+                var handlers = scope.Resolve<IEnumerable<IPartyEventHandler>>();
+
                 var originalDto = partySettingsDto.Clone();
                 var ctx = new PartySettingsUpdateCtx(this, partySettingsDto);
-                await handlers().RunEventHandler(h => h.OnUpdatingSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdatingSettings", ex));
+
+                await handlers.RunEventHandler(h => h.OnUpdatingSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdatingSettings", ex));
 
                 if (!ctx.ApplyChanges)
                 {
@@ -365,7 +369,7 @@ namespace Stormancer.Server.Plugins.Party
                 }
 
                 Log(LogLevel.Trace, "UpdateSettings", "Settings update accepted", partySettingsDto);
-                await handlers().RunEventHandler(h => h.OnUpdateSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdateSettings", ex));
+                await handlers.RunEventHandler(h => h.OnUpdateSettings(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnUpdateSettings", ex));
 
                 // If the event handlers have modified the settings, we need to notify the leader to invalidate their local copy.
                 // Make an additional bump to the version number to achieve this.
@@ -389,13 +393,10 @@ namespace Stormancer.Server.Plugins.Party
 
                 Dictionary<PartyMember, PartySettingsUpdateDto> updates = _partyState.PartyMembers.Values.ToDictionary(m => m, _ => new PartySettingsUpdateDto(_partyState));
 
-                await handlers().RunEventHandler(h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, updates)),
+                await handlers.RunEventHandler(h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, updates)),
                 ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnSendingSettingsToMember", ex));
 
-                await BroadcastStateUpdateRpc(
-                    PartySettingsUpdateDto.Route,
-                    updates
-                    );
+                await BroadcastStateUpdateRpc(PartySettingsUpdateDto.Route, updates);
             });
         }
 
@@ -417,7 +418,7 @@ namespace Stormancer.Server.Plugins.Party
                 throw new ArgumentNullException(nameof(partyUserStatus));
             }
 
-            var handlers = _handlers;
+            var handlers = _handlers();
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 if (ct.IsCancellationRequested)
@@ -448,7 +449,7 @@ namespace Stormancer.Server.Plugins.Party
                 await BroadcastStateUpdateRpc(BatchStatusUpdate.Route, update);
 
                 var eventHandlerCtx = new PlayerReadyStateContext(this, user);
-                await handlers().RunEventHandler(h => h.OnPlayerReadyStateChanged(eventHandlerCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnPlayerReadyStateChanged", ex));
+                await handlers.RunEventHandler(h => h.OnPlayerReadyStateChanged(eventHandlerCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnPlayerReadyStateChanged", ex));
 
                 bool shouldLaunchGameFinderRequest = false;
                 switch (eventHandlerCtx.GameFinderPolicy)
@@ -725,7 +726,7 @@ namespace Stormancer.Server.Plugins.Party
 
         public async Task SendPartyState(string recipientUserId, CancellationToken ct)
         {
-            var handlers = _handlers;
+            var handlers = _handlers();
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 if (ct.IsCancellationRequested)
@@ -758,7 +759,7 @@ namespace Stormancer.Server.Plugins.Party
 
         public async Task SendPartyStateAsRequestAnswer(RequestContext<IScenePeerClient> ctx)
         {
-            var handlers = _handlers;
+            var handlers = _handlers();
             await _partyState.TaskQueue.PushWork(async () =>
             {
                 if (ctx.CancellationToken.IsCancellationRequested)
@@ -779,7 +780,7 @@ namespace Stormancer.Server.Plugins.Party
             });
         }
 
-        private async Task<PartyStateDto> MakePartyStateDto(PartyMember recipient, Func<IEnumerable<IPartyEventHandler>> handlers)
+        private async Task<PartyStateDto> MakePartyStateDto(PartyMember recipient, IEnumerable<IPartyEventHandler> handlers)
         {
             var dto = new PartyStateDto
             {
@@ -790,9 +791,16 @@ namespace Stormancer.Server.Plugins.Party
                 Version = _partyState.VersionNumber
             };
 
-            await handlers().RunEventHandler(
-                h => h.OnSendingSettingsUpdateToMembers(new PartySettingsMemberUpdateCtx(this, new Dictionary<PartyMember, PartySettingsUpdateDto> { { recipient, dto.Settings } })),
-                ex => Log(LogLevel.Error, "MakePartyStateDto", "An exception was thrown by a OnSendingSettingsUpdateToMembers handler"));
+            await handlers.RunEventHandler(
+                h => h.OnSendingSettingsUpdateToMembers(
+                    new PartySettingsMemberUpdateCtx(
+                        this,
+                        new Dictionary<PartyMember,
+                        PartySettingsUpdateDto> { { recipient, dto.Settings } }
+                    )
+                ),
+                ex => Log(LogLevel.Error, "MakePartyStateDto", "An exception was thrown by a OnSendingSettingsUpdateToMembers handler")
+            );
 
             return dto;
         }